moved concanate_extension to core/mrio_system
